+++ builds/wtf_wikipedia.mjs
@@ -2302,6 +2302,15 @@ function preProcess(wiki) {
   // some html escaping
   wiki = wiki.replace(/&nbsp;/g, ' ');
   wiki = wiki.replace(/&ndash;/g, '–');
+  wiki = wiki.replace(/&mdash;/g, '—');
+  wiki = wiki.replace(/&amp;/g, '&');
+  wiki = wiki.replace(/&quot;/g, '"');
+  wiki = wiki.replace(/&apos;/g, "'");
+  wiki = wiki.replace(/&copy;/g, '©');
+  wiki = wiki.replace(/&reg;/g, '®');
+  wiki = wiki.replace(/&trade;/g, '™');
+  // wiki = wiki.replace(/&lt;/g, '<')
+  // wiki = wiki.replace(/&gt;/g, '>')
 
   //give it the inglorious send-off it deserves..
   wiki = kill_xml(wiki);
@@ -2611,7 +2620,7 @@ const formatting = function (obj) {
   return obj
 };
 
-const isNumber = /^[0-9,.]+$/;
+const isNumber = /^-?[0-9,.]+$/;
 
 const defaults$7 = {
   text: true,
@@ -4081,7 +4090,6 @@ Object.keys(methods$3).forEach((k) => {
 const list_reg = /^[#*:;|]+/;
 const bullet_reg = /^\*+[^:,|]{4}/;
 const number_reg = /^ ?#[^:,|]{4}/;
-const has_word = /[\p{Letter}_0-9\]}]/iu;
 
 // does it start with a bullet point or something?
 const isList = function (line) {
@@ -4117,7 +4125,7 @@ const grabList = function (lines, i) {
       break
     }
   }
-  sub = sub.filter((a) => a && has_word.test(a));
+  // sub = sub.filter((a) => a && has_word.test(a))
   sub = cleanList(sub);
   return sub
 };
@@ -4131,8 +4139,9 @@ const parseList = function (paragraph) {
     if (isList(lines[i])) {
       let sub = grabList(lines, i);
       if (sub.length > 0) {
-        lists.push(sub);
         i += sub.length - 1;
+        sub = sub.filter((s) => s.data.text !== '');
+        lists.push(sub);
       }
     } else {
       theRest.push(lines[i]);
@@ -9403,6 +9412,43 @@ const toJson = function (infobox, options) {
   return json
 };
 
+var latLngs = [
+  ['latitude', 'longitude'], //english
+  ['lat', 'lng'],
+  ['breitengrad', 'längengrad'], //german
+  ['latitud', 'longitud'], //spanish
+  ['zeměpisná šířka', 'zeměpisná délka'], //czech
+  ['шир', 'долг'], //russian
+];
+
+// lat_d: { text: '43', number: 43 },
+// lat_m: { text: '42', number: 42 },
+// lat_s: { text: '59', number: 59 },
+// lat_ns: { text: 'S' },
+// long_d: { text: '79', number: 79 },
+// long_m: { text: '20', number: 20 },
+// long_s: { text: '26', number: 26 },
+// long_ew: { text: 'Z' },
+
+
+// latd: { text: '43', number: 43 },
+// latm: { text: '42', number: 42 },
+// lats: { text: '0', number: 0 },
+// latns: { text: 'N' },
+// longd: { text: '79', number: 79 },
+// longm: { text: '24', number: 24 },
+// longs: { text: '0', number: 0 },
+// longew: { text: 'W' },
+
+// gjer_g: { text: '43', number: 43 },
+// gjer_m: { text: '42', number: 42 },
+// gjer_s: { text: '0', number: 0 },
+// gjer_dr: { text: 'N' },
+// gjat_g: { text: '79', number: 79 },
+// gjat_m: { text: '24', number: 24 },
+// gjat_s: { text: '0', number: 0 },
+// gjat_dr: { text: 'W' },
+
 const normalize = (str = '') => {
   str = str.toLowerCase();
   str = str.replace(/[-_]/g, ' ');
@@ -9497,6 +9543,18 @@ const methods$2 = {
       }
       return h
     }, {})
+  },
+  coordinates: function () {
+    // loop through i18n latitude and longitude properties
+    for (let i = 0; i < latLngs.length; i += 1) {
+      let a = latLngs[i];
+      let lat = this.get(a[0])?.json()?.number;
+      let lng = this.get(a[1])?.json()?.number;
+      if (lat && lng) {
+        return { template: 'infobox/lat-long', lat, lng }
+      }
+    }
+    return null
   },
 };
 //aliases
@@ -9721,7 +9779,7 @@ const parseRefs = function (section) {
   let references = [];
   let wiki = section._wiki;
 
-  wiki = wiki.replace(/ ?<ref>([\s\S]{0,1800}?)<\/ref> ?/gi, function (all, txt) {
+  wiki = wiki.replace(/ ?<ref>([\s\S]{0,4000}?)<\/ref> ?/gi, function (all, txt) {
     let found = false;
     // there could be more than 1 template inside a <ref><ref>
     let arr = findFlat(txt);
@@ -10271,7 +10329,13 @@ class Section {
    */
   coordinates() {
     let arr = [...this.templates('coord'), ...this.templates('coor')];
-    return arr.map((tmpl) => tmpl.json())
+    let list = arr.map((tmpl) => tmpl.json());
+    //try to get coord from infoboxes
+    let inf = this.infoboxes()[0];
+    if (inf && inf.coordinates()) {
+      list.push(inf.coordinates());
+    }
+    return list
   }
 
   /**
@@ -11299,29 +11363,14 @@ Document.prototype.redirects = Document.prototype.redirectTo;
  * @private
  * @param {Array} res
  * @param {string | number | Array<number> | Array<string>} title
- * @returns {null| Document | Document[]} null if there are no results or Document if there is one responses and Document array if there are multiple responses
+ * @returns {null | Document | Document[]} `Document | null` if `title` is scalar, `or Document[]` if `title` is an array
  */
 const parseDoc = function (res, title) {
-  res = res || [];
-  // filter out undefined
-  res = res.filter((o) => o);
+  const results = (res ?? [])
+    .filter((o) => o != null)
+    .map(o => new Document(o.wiki, o.meta));
 
-  // put all the responses into Document formats
-  let docs = res.map((o) => {
-    return new Document(o.wiki, o.meta)
-  });
-
-  // if the list is empty than there are no results
-  if (docs.length === 0) {
-    return null
-  }
-
-  // if there is only one response then we can get it out of the array
-  if (!isArray(title) && docs.length === 1) {
-    return docs[0]
-  }
-
-  return docs
+  return isArray(title) ? results : results[0] ?? null
 };
 
 /**
@@ -11355,7 +11404,6 @@ const makeHeaders = function (options) {
   }
 };
 
-/* eslint-disable no-console */
 const isUrl = /^https?:\/\//;
 
 /**
@@ -11405,6 +11453,9 @@ const fetch = function (title, options, callback) {
   if (typeof options === 'string') {
     options = { lang: options };
   }
+  if (typeof title.href === 'string') {
+    title = title.href;
+  }
   options = { ...defaults, ...options };
   options.title = title;
 
@@ -11415,29 +11466,26 @@ const fetch = function (title, options, callback) {
   const url = makeUrl(options);
   const headers = makeHeaders(options);
 
-  return unfetch(url, headers)
+  const promise = unfetch(url, headers)
     .then((res) => res.json())
     .then((res) => {
       if (!res) {
         throw new Error(`No JSON Data Found For ${url}`)
       }
-      let data = getResult(res, options);
-      data = parseDoc(data, title);
-      if (callback) {
+      const result = getResult(res, options);
+      const data = parseDoc(result, title);
+      if (typeof callback === 'function') {
         callback(null, data);
       }
       return data
-    })
-    .catch((e) => {
-      console.error(e);
-      if (callback) {
-        callback(e, null);
-      }
-      return null
-    })
+    });
+
+  return typeof callback === 'function'
+    ? promise.catch((e) => callback(e, null))
+    : promise
 };
 
-var version = '10.4.0';
+var version = '10.4.1';
 
 /* eslint-disable no-console */