+++ src/02-section/Section.js
@@ -1,6 +1,5 @@
 import toJSON from './toJson.js'
 import setDefaults from '../_lib/setDefaults.js'
-// import parseHeading from './heading.js'
 import parseTable from '../table/index.js'
 import parseParagraphs from '../03-paragraph/index.js'
 import parseTemplates from '../template/index.js'
@@ -225,7 +224,13 @@ class Section {
    */
   coordinates() {
     let arr = [...this.templates('coord'), ...this.templates('coor')]
-    return arr.map((tmpl) => tmpl.json())
+    let list = arr.map((tmpl) => tmpl.json())
+    //try to get coord from infoboxes
+    let inf = this.infoboxes()[0]
+    if (inf && inf.coordinates()) {
+      list.push(inf.coordinates())
+    }
+    return list
   }
 
   /**