+++ src/list/index.js
@@ -39,7 +39,7 @@ const grabList = function (lines, i) {
       break
     }
   }
-  sub = sub.filter((a) => a && has_word.test(a))
+  // sub = sub.filter((a) => a && has_word.test(a))
   sub = cleanList(sub)
   return sub
 }
@@ -53,8 +53,9 @@ const parseList = function (paragraph) {
     if (isList(lines[i])) {
       let sub = grabList(lines, i)
       if (sub.length > 0) {
-        lists.push(sub)
         i += sub.length - 1
+        sub = sub.filter((s) => s.data.text !== '')
+        lists.push(sub)
       }
     } else {
       theRest.push(lines[i])