+++ tests/integration/lists.test.js
@@ -98,3 +98,17 @@ test('ordered list', function (t) {
   t.equal(doc.text(), '1. first item\n\n2. second item\n\n3. third item', 'text 4')
   t.end()
 })
+
+test('empty lists', function (t) {
+  let str = `
+*{{cite news|title=Love—In Other Words}}
+*{{cite news|title=Christmas to Me}}
+*{{cite news|title=When Children Discover America}}
+*{{cite news|title=Romance and High Adventure}} Show Just once.
+*{{cite news|title=Open letter to Oprah Winfrey}} 
+`
+  let doc = wtf(str)
+  t.equal(doc.lists().length, 1, 'got 1 list')
+  t.equal(doc.text(), '\n * Show Just once.', 'list text output')
+  t.end()
+})